+++ Assets/Scripts/view/GameObject/ZoneHumbleObject.cs
@@ -19,7 +19,7 @@ public class ZoneHumbleObject : MonoBehaviour
     private const float TangnetApproximation = 0.001f;
     public Zone Zone { get; private set; }
 
-    public void Init(ulong id, SplineContainer splineContainer)
+    public void Init(uint id, SplineContainer splineContainer)
     {
         Assert.AreEqual(1, splineContainer.Splines.Count());
